backend/requirements.txt
discord.py
flask
flask-cors
python-dotenv